Added the ability to pass the units and symmetry info in the field map
Added some new information to the header for the field maps to indicate
the units and if the field map is symmetric.

The new header now can include the units for the position and field as
strings (e.g. cm and tesla or V/cm) after the origin offset and grid size.

The function reading the file will ask the GEANT UnitTable for the
definition of the units and apply the correct factor. This does require
the unit to be defined in the unit table otherwise this will crash.

The last entry on the header line is a boolean flag (i.e. true or false)
which indicates if the field map is symmetric about XYZ. This allows the
field to be defined using only the upper-octant (i.e. only the positive
XYZ coordinates).