Modified point calc in ArbElec/ArbMagField to account for symmetry
To properly account for symmetry when the field map is given as the
upper-octant the coordinate offset is subtracted first before being
passed to the interpolation function.

The offset values passed to the interpolation function are now all
zeroes since they have already been used (and to reuse the same function
call).